ttools: use generics throughout
Eliminate or suppress all compile-time warnings at java 8.
There are a lot of lines of source code changed, but most of them
do not result in any changes that don't get erased at compile time.
There should be no behaviour changes.

SuppressWarnings is used in a few places, notably when dealing with
classes like JComboBox that have acquired generics since java 6.

+++ ttools/src/main/uk/ac/starlink/ttools/convert/Iso8601ToDecimalYear.java
@@ -50,7 +50,7 @@ public class Iso8601ToDecimalYear extends Iso8601Converter {
         long y0 = cal.getTimeInMillis();
         cal.set( Calendar.YEAR, year + 1 );
         long y1 = cal.getTimeInMillis();
-        long t = y0 + (long) Math.round( frac * ( y1 - y0 ) );
+        long t = y0 + Math.round( frac * ( y1 - y0 ) );
         return unixMillisToMjd( t );
     }
